make it work by hacking libmp4v2
libmp4v2 provides custom io operations, but still uses a non customizable stat
to get the file size. The size is hardcoded for now. Proper solution would be
to add getSize() to MP4FileProvider.

The rest of the changes are just cleaning up and enabling optimization which
for some reason needs DISABLE_EXCEPTION_CATCHING=0.