properly fix problem in libmp4v2 to get rid of hack
A new function filesize() was added to MP4FileProvider and classes using this
structure. This way it's now possible to read the mp4 file from a buffer in
memory by providing custom io functions and also get the file size right.
Previously it was always using stat() on the file name which isn't working with
a buffer..