read 3 MB instead of 2 MB from the head of the mp4 file
For some videos reading 2 MB is not enough. I'm not sure what the right size
is. Using a hardcoded size is bad anyway.

A proper solution would be to read more when libmp4v2 wants to read more. We
can't have the whole file in memory. The problem is: Javascript file API is
asynchronous.