added mp4check which extracts bitrate and codecs
mp4check takes a look at a video file and checks the bitrate and audio and
video codecs in it. This is quite useful for video portals to decide if this
video file has the proper settings and is worth to be uploaded.

At the moment it just ckecks for a MPEG-4 AAC stream compatible with the flash
player and a H264 video stream.