added nicer interface to mp4check
Hide the ugly details of calling compiled functions by providing a nicer class
interface. The examples mp4check.xml and mp4checkminimal.xml are changed to use
the new interface. Makes usage way easier.

The javascript class in interface.js is added to the end of the compiled js
file after emscripten is run. It's not using emcc's --post-js option, because
the closure compiler would mess around with it.